+++ internal/authentication/token_repository.go
@@ -15,6 +15,7 @@ const (
 
 type Token struct {
 	Id           string    `json:"id"`
+	TenantID     string    `json:"tenantId"`
 	Email        string    `json:"email"`
 	ClientId     string    `json:"clientId"`
 	AccessToken  string    `json:"accessToken"`
@@ -24,10 +25,10 @@ type Token struct {
 }
 
 type TokenRepository interface {
-	Create(ctx context.Context, email, clientId, accessToken, refreshToken string) error
-	Delete(ctx context.Context, email, clientId string) error
-	DeleteByEmail(ctx context.Context, email string) error
-	Read(ctx context.Context, email, clientId string) (*Token, error)
+	Create(ctx context.Context, tenantID, email, clientId, accessToken, refreshToken string) error
+	Delete(ctx context.Context, tenantID, email, clientId string) error
+	DeleteByEmail(ctx context.Context, tenantID, email string) error
+	Read(ctx context.Context, tenantID, email, clientId string) (*Token, error)
 }
 
 type DefaultTokenRepository struct {
@@ -38,12 +39,13 @@ func NewDefaultTokenRepository(db *mongo.Database) *DefaultTokenRepository {
 	return &DefaultTokenRepository{db}
 }
 
-func (t *DefaultTokenRepository) Create(ctx context.Context, email, clientId, accessToken, refreshToken string) error {
+func (t *DefaultTokenRepository) Create(ctx context.Context, tenantID, email, clientId, accessToken, refreshToken string) error {
 	collection := t.db.Collection(collectionTokens)
 
-	filter := bson.D{{"email", email}, {"clientId", clientId}}
+	filter := bson.M{"tenantId": tenantID, "email": email, "clientId": clientId}
 	update := bson.D{
 		{"$set", bson.D{
+			{"tenantId", tenantID},
 			{"accessToken", accessToken},
 			{"refreshToken", refreshToken},
 			{"modifiedAt", time.Now()},
@@ -62,28 +64,28 @@ func (t *DefaultTokenRepository) Create(ctx context.Context, email, clientId, ac
 	return nil
 }
 
-func (t *DefaultTokenRepository) DeleteByEmail(ctx context.Context, email string) error {
+func (t *DefaultTokenRepository) DeleteByEmail(ctx context.Context, tenantID, email string) error {
 	collection := t.db.Collection(collectionTokens)
-	_, err := collection.DeleteOne(ctx, bson.M{"email": email})
+	_, err := collection.DeleteOne(ctx, bson.M{"tenantId": tenantID, "email": email})
 	if err != nil {
 		return err
 	}
 	return nil
 }
 
-func (t *DefaultTokenRepository) Delete(ctx context.Context, email, clientId string) error {
+func (t *DefaultTokenRepository) Delete(ctx context.Context, tenantID, email, clientId string) error {
 	collection := t.db.Collection(collectionTokens)
-	_, err := collection.DeleteOne(ctx, bson.M{"email": email, "clientId": clientId})
+	_, err := collection.DeleteOne(ctx, bson.M{"tenantId": tenantID, "email": email, "clientId": clientId})
 	if err != nil {
 		return err
 	}
 	return nil
 }
 
-func (t *DefaultTokenRepository) Read(ctx context.Context, email, clientId string) (*Token, error) {
+func (t *DefaultTokenRepository) Read(ctx context.Context, tenantID, email, clientId string) (*Token, error) {
 	collection := t.db.Collection(collectionTokens)
 	var token Token
-	err := collection.FindOne(ctx, bson.M{"email": email}).Decode(&token)
+	err := collection.FindOne(ctx, bson.M{"tenantId": tenantID, "email": email}).Decode(&token)
 	if err != nil {
 		return nil, err
 	}
